Ignore Files, Swap Files and Text File
The file .gitignore lets us ignore files that may be needed for our
project but that are either generated or otherwise not needed to be
public (at least public to the repository).

The ignored text.txt file is due to two uses, one for exercise dealing
with document statistics and the other is string replacement.  The same
filename is often used for both, specifically 'text.txt'.  But we also
require that the programs don't hardcode the name.

The ignore file now has the vim swp files ignored as well.

This commit (in bad form) also has the document stats program added in.